requirements.txt
asgiref==3.5.2
autopep8==1.6.0
beautifulsoup4==4.11.1
certifi==2022.5.18.1
cloudinary==1.29.0
config==0.5.1
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==4.0.2
django-bootstrap3==21.2
django-bootstrap5==21.3
django-braces==1.15.0
django-cors-headers==3.13.0
django-heroku==0.3.1
django-jquery==3.1.0
django-model-utils==4.2.0
django-registration==3.3
django-star-ratings==0.9.2
django-voting==1.0
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.1.1
psycopg2==2.9.3
pycodestyle==2.8.0
PyJWT==1.7.1
python-decouple==3.6
pytz==2022.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
swapper==1.3.0
toml==0.10.2
urllib3==1.26.9
whitenoise==6.2.0
